app/api/views.py
from rest_framework import viewsets
from core.models import Item
from core.serializers import ItemSerializer


class ItemViewSet(viewsets.ModelViewSet):
    queryset = Item.objects.all()
    serializer_class = ItemSerializer

